+++ src/main/java/bio/terra/common/opentelemetry/OpenTelemetryConfig.java
@@ -1,4 +1,4 @@
-package bio.terra.common.prometheus;
+package bio.terra.common.opentelemetry;
 
 import io.opentelemetry.exporter.prometheus.PrometheusHttpServer;
 import io.opentelemetry.instrumentation.spring.autoconfigure.EnableOpenTelemetry;
@@ -6,31 +6,19 @@ import io.opentelemetry.sdk.metrics.InstrumentSelector;
 import io.opentelemetry.sdk.metrics.SdkMeterProvider;
 import io.opentelemetry.sdk.metrics.View;
 import io.opentelemetry.sdk.resources.Resource;
-import io.prometheus.client.exporter.HTTPServer;
 import java.util.Optional;
 import org.springframework.beans.factory.ObjectProvider;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Primary;
 import org.springframework.data.util.Pair;
 
-/** Spring Beans that exposes any OpenTelemetry metrics via a Prometheus {@link HTTPServer}. */
 @Configuration
 @EnableOpenTelemetry
-public class PrometheusComponent {
-  @Bean(destroyMethod = "close")
-  @ConditionalOnProperty(
-      name = "terra.common.prometheus.endpointEnabled",
-      havingValue = "true",
-      matchIfMissing = true)
-  public PrometheusHttpServer prometheusHttpServer(PrometheusProperties prometheusProperties) {
-    return PrometheusHttpServer.builder().setPort(prometheusProperties.getEndpointPort()).build();
-  }
-
+public class OpenTelemetryConfig {
   @Bean
   @Primary
-  public SdkMeterProvider prometheusMeterProvider(
+  public SdkMeterProvider terraMeterProvider(
       Resource otelResource,
       Optional<PrometheusHttpServer> prometheusHttpServer,
       ObjectProvider<Pair<InstrumentSelector, View>> views) {
