+++ src/test/java/bio/terra/common/tracing/TracingTestApplication.java
@@ -12,7 +12,7 @@ import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
     },
     // Scan the common tracing package to pick up tracing auto configuration. Scan the logging
     // package to get request ID setting.
-    scanBasePackages = {"bio.terra.common.tracing", "bio.terra.common.logging"})
+    scanBasePackages = {"bio.terra.common.logging"})
 public class TracingTestApplication {
   // N.B. this main method isn't called by the test! This is a demonstration of how a Terra
   // application might initialize the tracing setup at the very beginning of the Spring Boot flow.
