+++ src/test/resources/application-tracing-test.properties
@@ -1,4 +1,1 @@
-terra.common.tracing.samplingRate: 1.0
-# Empty list of urlPatterns to allow all URLs to be traced.
-terra.common.tracing.urlPatterns:
 terra.common.tracing.stackdriverExportEnabled: false
